% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Implements agglomerative hierarchical clustering}
\usage{
hier_clust(dat, k)
}
\arguments{
\item{dat}{A data frame}

\item{k}{The number of clusters you want}
}
\value{
A data frame with the cluster assignments
}
\description{
Implements agglomerative hierarchical clustering
}
