% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Implements a basic k-means algorithm}
\usage{
k_means(dat, k, pca = FALSE)
}
\arguments{
\item{dat}{A data frame}

\item{k}{The number of observations to use as starting points}

\item{pca}{An option if you want to perform PCA before doing the k_means clustering}
}
\value{
A list with the cluster assignments and total sum of squares
}
\description{
Implements a basic k-means algorithm
}
